add perm and c

+++ pointers_arrays_strings/8-print_diagsums.c
@@ -2,26 +2,25 @@
 #include <stdio.h>
 
 /**
- * print_diagsums - Affiche la somme des deux diagonales d'une matrice carrée.
- * @a: Pointeur vers le premier élément de la matrice (vu comme un tableau 1D).
- * @size: La taille d'un côté de la matrice (ex: 3 pour une matrice 3x3).
+ * print_diagsums - Affiche la somme des deux diagonales
+ * d'une matrice carrée d'entiers.
+ * @a: Pointeur vers le premier élément de la matrice.
+ * @size: Taille d'un côté de la matrice.
  */
 void print_diagsums(int *a, int size)
 {
 	int i;
-	int sum1 = 0;
-	int sum2 = 0;
+	long int sum1 = 0;
+	long int sum2 = 0;
 
 	for (i = 0; i < size; i++)
 	{
-		/* Diagonale principale : index (ligne * size) + ligne */
-		/* Ce qui revient à i * (size + 1) */
+		/* Diagonale principale (0, 4, 8... pour une matrice 3x3) */
 		sum1 += a[i * (size + 1)];
 
-		/* Diagonale secondaire : index (ligne * size) + (size - 1 - ligne) */
-		/* Ce qui revient à (i + 1) * (size - 1) */
+		/* Diagonale secondaire (2, 4, 6... pour une matrice 3x3) */
 		sum2 += a[(i + 1) * (size - 1)];
 	}
 
-	printf("%d, %d\n", sum1, sum2);
+	printf("%ld, %ld\n", sum1, sum2);
 }
